backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
geopy==2.4.1
pytz==2024.1
bcrypt==4.2.0
PyJWT==2.9.0
